requirements.txt
Django~=3.0.7
Pillow~=7.1.12
django-admin-sortable2~=0.7.6